Fixes for Pi compatibility.
Previous build was working with macOS but refused to work with my Pi.

Looks like removing the lines in tusb_config.h was the magic fix, but
I haven't delved into why.

Additionally I have added an __isr for feeding audio, since this (in
theory) should provide a new audio buffer to the pool every time an
old one is consumed.

Finally I've migrated some of our tweaks to the descriptors back into
this new code, including specifying a fixed clock and ensuring the
right string IDs (mostly) are used.